Address code-review comments.
